refactor: unify service directory structures
