+++ services/asr-python/config/config.json
