+++ services/asr-python/src/services/providers/openai_whisper.py
