+++ services/chat-ai-python/src/utils/example.py
