+++ services/gateway-python/config/config.json
