+++ services/memory-python/config/config.json
